% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{AWR.KMS-package}
\alias{AWR.KMS-package}
\title{An R client to Amazon Key Management Service}
\description{
This is a simple wrapper around the most important features of the related Java SDK.
}
\references{
\url{http://docs.aws.amazon.com/AWSJavaSDK/latest/javadoc/com/amazonaws/services/kms/AWSKMSClient.html}
}
