% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kms.R
\name{kms_decrypt}
\alias{kms_decrypt}
\title{Decrypt cipher into plain text via KMS}
\usage{
kms_decrypt(cipher, return = c("string", "raw"))
}
\arguments{
\item{cipher}{Base64-encoded ciphertext}

\item{return}{return format}
}
\value{
decrypted text as string or raw
}
\description{
Decrypt cipher into plain text via KMS
}
\references{
\url{http://docs.aws.amazon.com/AWSJavaSDK/latest/javadoc/com/amazonaws/services/kms/AWSKMSClient.html#decrypt-com.amazonaws.services.kms.model.DecryptRequest-}
}
\seealso{
kms_encrypt
}
