% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kms.R
\name{kms_decrypt_file}
\alias{kms_decrypt_file}
\title{Decrypt file via KMS}
\usage{
kms_decrypt_file(file)
}
\arguments{
\item{file}{base file path (without the \code{enc} or \code{key} suffix)}
}
\value{
decrypted file path
}
\description{
Decrypt file via KMS
}
\seealso{
kms_encrypt kms_encrypt_file
}
