% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{kms_init}
\alias{kms_init}
\title{Initialize a KMS Client}
\usage{
kms_init()
}
\value{
the KMS Client is cached within the package namespace as \code{kms_client}
}
\description{
Initialize a KMS Client
}
\note{
This is automatically called when the KMS Client is required and you probably will never have to run this manually, except if you have a very non-standard workflow.
}
\keyword{internal}
