% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kms.R
\name{retry}
\alias{retry}
\title{Retry the query up to 3 times on failure}
\usage{
retry(cmd, retries = 0)
}
\arguments{
\item{cmd}{expression}

\item{retries}{number of previous retries}
}
\description{
Retry the query up to 3 times on failure
}
\keyword{internal}
